DDD.jl

DDD.AbstractCrystalStruct

DDD.AbstractDistribution

DDD.AbstractDlnSeg

DDD.AbstractDlnStr

DDD.AbstractIntegrator

DDD.AbstractMesh

DDD.AbstractMobility

DDD.AbstractShapeFunction

DDD.DislocationLoop

DDD.DislocationNetwork

DDD.DislocationP

DDD.IntegrationP

DDD.MaterialP

DDD.SlipSystem

DDD.nodeType

DDD.calcSegSegForce

DDD.calcSelfForce

DDD.checkNetwork

DDD.compStruct

DDD.coordIdx

DDD.coordLbl

DDD.dataCond

DDD.extAngle

DDD.hatStress

DDD.idxCond

DDD.idxLabel

DDD.inclusiveComparison

DDD.intAngle

DDD.limits!

DDD.load

DDD.loadDislocationLoop

DDD.loadDislocationP

DDD.loadIntegrationP

DDD.loadMaterialP

DDD.loadNetwork

DDD.loadParams

DDD.loadSlipSystem

DDD.loopDistribution

DDD.makeConnect!

DDD.makeInstanceDict

DDD.makeLoop

DDD.makeNetwork

DDD.makeNetwork!

DDD.makeSegment

DDD.makeTypeDict

DDD.pkForce

DDD.rot3D

DDD.save

DDD.shapeFunction

DDD.shapeFunctionDeriv

DDD.subTypeTree

DDD.translateEnum

DDD.translatePoints

JSON.Writer.lower

DDD.AbstractDistribution — Type

abstract type AbstractDistribution end
struct Zeros <: AbstractDistribution end
struct Rand <: AbstractDistribution end
struct Randn <: AbstractDistribution end
struct Regular <: AbstractDistribution end

Distributions for dislocation sources.

DDD.AbstractDlnStr — Type

Dislocation structures have different classifications. Prismatic loops are made up only of edge segments with the same slip system; shear loops are made up of a mixture of segment types with the same slip system; jogs and kinks are steps not contained in the slip plane.

abstract type AbstractDlnStr end
struct loopDln <: AbstractDlnStr end    # Unclassified loop.
struct loopPrism <: AbstractDlnStr end  # Prismatic loop.
struct loopShear <: AbstractDlnStr end  # Shear loop.
struct loopJog <: AbstractDlnStr end    # Jog.
struct loopKink <: AbstractDlnStr end   # Kink.

DDD.DislocationLoop — Type

struct DislocationLoop{
    T1 <: AbstractDlnStr,
    T2 <: Int64,
    T3 <: Union{T where {T <: Float64}, AbstractArray{<:Float64, N} where {N}},
    T4 <: Union{T where {T <: Int64}, AbstractArray{<:Int64, N} where {N}},
    T5 <: AbstractArray{<:Int64, N} where {N},
    T6 <: AbstractArray{<:Float64, N} where {N},
    T7 <: Vector{<:nodeType},
    T8 <: Float64,
    T9 <: AbstractDistribution,
}

    loopType::T1
    numSides::T2
    nodeSide::T2
    numLoops::T2
    segLen::T3
    slipSystem::T4
    links::T5
    slipPlane::T6
    bVec::T6
    coord::T6
    label::T7
    buffer::T8
    range::T6
    dist::T9

Dislocation loop structure generated via the constructor makeLoop.

DDD.DislocationNetwork — Type

DislocationNetwork{
    T1 <: AbstractArray{<:Int64, N} where {N},
    T2 <: AbstractArray{<:Float64, N} where {N},
    T3 <: Vector{nodeType},
    T4 <: Int64,
    T5 <: Integer,
}
    links::T1
    slipPlane::T2
    bVec::T2
    coord::T2
    label::T3
    numNode::T4 = 0     # Total number of nodes in network.
    numSeg::T4 = 0      # Total number of segs in network.
    maxConnect::T5 = 4  # Maximum connectivity of nodes.
    connectivity::T1
    linksConnect::T1
    segIdx::T1          # segIdx[:,1] is the segment index. Used to find the bVec and slipPlane of a real segment. segIdx[:,2:3] are the indices of the nodes involved in a given link, used to find their coordinates.

Dislocation Network structure. See DislocationLoop, makeNetwork and makeNetwork! for further details.

DDD.SlipSystem — Type

struct SlipSystem{
    T1 <: AbstractString,
    T2 <: AbstractArray{<:Float64, N} where {N},
}
    name::T1
    slipPlane::T2
    bVec::T2

Slip systems.

DDD.nodeType — Type

Dislocation nodes have labels that change how they are treated by the simulation. There are only given types of nodes so these labels may only take on predefined values and error for anything else.

@enum nodeType begin
    undef = -1  # Undefined node, value at initialisation.
    intMob = 0  # Internal mobile node.
    intFix = 1  # Internal fixed node.
    srfMob = 2  # Mobile surface node.
    srfFix = 3  # Fixed surface node.
    ext = 4     # External node.
end

DDD.calcSegSegForce — Method

!!! Note This function is based on the SegSegForces function by A. Arsenlis et al. It is optimised for speed and reusability.

It implements the analytical solution of the force between two dislocation segments. Details are found in Appendix A.1. in "Enabling Strain Hardening Simulations with Dislocation Dynamics" by A. Arsenlis et al.

At a high level this works by creating a local coordinate frame using the line directions of the dislocation segments and a vector orthogonal to them. The line integrals are then evaluated parametrically utilising this local coordinate. BibTex citation here:

@article{Arsenlis_2007, doi = {10.1088/0965-0393/15/6/001}, url = {https://doi.org/10.1088%2F0965-0393%2F15%2F6%2F001}, year = 2007, month = {jul}, publisher = {{IOP} Publishing}, volume = {15}, number = {6}, pages = {553–595}, author = {A Arsenlis and W Cai and M Tang and M Rhee and T Oppelstrup and G Hommes and T G Pierce and V V Bulatov}, title = {Enabling strain hardening simulations with dislocation dynamics}, journal = {Modelling and Simulation in Materials Science and Engineering}, abstract = {Numerical algorithms for discrete dislocation dynamics simulations are investigated for the purpose of enabling strain hardening simulations of single crystals on massively parallel computers. The algorithms investigated include the calculation of forces, the equations of motion, time integration, adaptive mesh refinement, the treatment of dislocation core reactions and the dynamic distribution of data and work on parallel computers. A simulation integrating all these algorithmic elements using the Parallel Dislocation Simulator (ParaDiS) code is performed to understand their behaviour in concert and to evaluate the overall numerical performance of dislocation dynamics simulations and their ability to accumulate percent of plastic strain.} }

DDD.calcSelfForce — Method

function calcSelfForce(
    dlnParams::DislocationP,
    matParams::MaterialP,
    network::DislocationNetwork,
)

Calculates the self-interaction force felt by two nodes in a segment. Naturally the forces are equal and opposite to each other.

DDD.checkNetwork — Method

checkNetwork(network::DislocationNetwork)

Checks the validity of the dislocation network. It ensures the following conditions are met by the member variables of network:

connectivity and links have the same number of non-zero entries;

all entries in bVec are non-zero;

only the trailing columns of connectivity are zeros;

consistency between connectivity and links;

bVec is conserved among connected nodes;

entries in links are unique;

consistency betwen connectivity and linksConnect

DDD.compStruct — Method

compStruct(arg1, arg2; verbose::Bool=false)

Function that compares values of the fields of two variables arg1 and arg2 with the same structure. If verbose = true, it will print which fields are different from each other.

Examples

julia> struct MyStruct1; x; end
julia> test1 = MyStruct1(1)
MyStruct1(1)
julia> test2 = MyStruct1(5)
MyStruct1(5)
julia> compStruct(test1, test2; verbose = true)
Structures differ in field: x.
false
julia> compStruct(1, 1; verbose = true)
true
julia> compStruct(1, [1]; verbose = true)
false

DDD.extAngle — Method

extAngle(n::Int64)

Calculates the exterior angle of a regular polygon with n sides.

DDD.intAngle — Method

intAngle(n::Int64)

Calculates the interior angle of a regular polygon with n sides.

DDD.load — Method

load(filename::AbstractString)

Wrapper for JSON.parsefile(filename).

DDD.loadDislocationLoop — Method

function loadDislocationLoop(
    dict::Dict{T1, T2} where {T1, T2},
    slipSystem::SlipSystem,
)

Loads initial dislocation structure out of a dictionary loaded from a JSON file. Returns a variable of type DislocationLoop.

DDD.loadDislocationP — Method

loadDislocationP(dict::Dict{T1, T2}) where {T1, T2}

Loads dislocation parameters out of a dictionary loaded from a JSON file. Returns a variable of type DislocationP.

DDD.loadIntegrationP — Method

loadIntegrationP(dict::Dict{T1, T2}) where {T1, T2}

Loads integration parameters out of a dictionary loaded from a JSON file. Returns a variable of type IntegrationP.

DDD.loadMaterialP — Method

loadMaterialP(dict::Dict{T1, T2}) where {T1, T2}

Loads material parameters out of a dictionary loaded from a JSON file. Returns a variable of type MaterialP.

DDD.loadNetwork — Method

loadNetwork(fileDislocationNetwork::AbstractString)

Loads a dislocation network from a JSON file. Returns a DislocationNetwork.

DDD.loadParams — Method

loadParams(
    fileDislocationP::AbstractString,
    fileMaterialP::AbstractString,
    fileIntegrationP::AbstractString,
    fileSlipSystem::AbstractString,
    fileDislocationLoop::AbstractString,
)

Loads simulation parameters out of a dictionary loaded from a JSON file. Returns a tuple of variable types (DislocationP, MaterialP, IntegrationP, SlipSystem, DislocationLoop) or vectors of those types.

DDD.loadSlipSystem — Method

loadSlipSystem(dict::Dict{T1, T2}) where {T1, T2}

Loads slip systems out of a dictionary loaded from a JSON file. Returns a variable of type SlipSystem.

DDD.loopDistribution — Method

loopDistribution(dist<:AbstractDistribution, n::Integer, args...; kw...)

Returns n points according to the concrete subtype of AbstractDistribution given. Overload this function with new concrete subtypes and custom distributions. This and limits! are used in translatePoints to distribute dislocations in the simulation domain.

DDD.makeNetwork — Function

makeNetwork(
    sources::Union{
            DislocationLoop,
            AbstractVector{<:DislocationLoop}
        }, # Dislocation structures.
    maxConnect::Integer = 4,
    memBuffer::Integer = 10, # Buffer for memory allocation. The code will allocate the total number of nodes times `memBuffer` to reduce dynamic memory allocation during runtime.
    args...;
    checkConsistency::Bool = false, # Check the consistency of the network.
    kw...,
)

Constructor for DislocationNetwork, see makeNetwork! for in-place version.

DDD.makeNetwork! — Function

makeNetwork!(
    network::DislocationNetwork,
    sources::Union{
        DislocationLoop,
        AbstractVector{<:DislocationLoop}
    },
    maxConnect::Integer = 4,
    args...;
    checkConsistency::Bool = false,
    kw...,
)

In-place constructor for DislocationNetwork, see makeNetwork for constructor.

DDD.makeTypeDict — Method

function makeTypeDict(valType::DataType)

Inputs contain strings that correspond to DDD data types. This function atuomatically creates a dictionary for all concrete subtypes of a given valType.

Examples

julia> abstract type MyAbstractType end
julia> struct MyStruct1 <: MyAbstractType end
julia> struct MyStruct2 <: MyAbstractType end
julia> makeTypeDict(MyAbstractType)
Dict{String,Any} with 4 entries:
  "DDD.MyStruct1()" => MyStruct1()
  "DDD.MyStruct2()" => MyStruct2()
  "MyStruct1()"     => MyStruct1()
  "MyStruct2()"     => MyStruct2()

DDD.rot3D — Method

rot3D(xyz::Vector{<:Real}, uvw::Vector{<:Real}, abc::Vector{<:Real},
    θ::Real)

Rotate point xyz about the line with direction vector uvw that crosses the point abc by the angle θ. Further details found here.

Examples

julia> rot3D([1;1;1],[1;0;0],[0;0;0],π/2)
3-element Array{Float64,1}:
  1.0
 -0.9999999999999999
  1.0
julia> rot3D([1;1;1],[1;0;0],[0;0;0],-π/2)
3-element Array{Float64,1}:
1.0
1.0
-0.9999999999999999
julia> rot3D([1;1;1],[1;0;0],[0;0;0],π)
3-element Array{Float64,1}:
  1.0
 -1.0000000000000002
 -0.9999999999999999

DDD.save — Method

save(filename::AbstractString, args...; mode::AbstractString = "w")

Wrapper for JSON.print to a file, args are the variables or structures you want to save.

DDD.shapeFunction — Method

shapeFunction(
    shape<:AbstractShapeFunction,
    x::Union{Float64, AbstractVector{<:Float64}},
    y::Union{Float64, AbstractVector{<:Float64}},
    z::Union{Float64, AbstractVector{<:Float64}}

Returns the shape functions of type typeof(shape) <: AbstractShapeFunction. If x,y,z are floats returns a vector of length N, different shape functons have different numbers of nodes. If given vectors, returns an array of size (N, length(x)).

Note

All coordinate vectors must be of equal length.

shapeFunctionDeriv are the 1st order derivatives of the shape functions.

DDD.shapeFunctionDeriv — Method

shapeFunctionDeriv(
    shape<:AbstractShapeFunction,
    x::Union{Float64, AbstractVector{<:Float64}},
    y::Union{Float64, AbstractVector{<:Float64}},
    z::Union{Float64, AbstractVector{<:Float64}}

Returns the first order derivative of the shape functions, shapeFunction, of type typeof(shape) <: AbstractShapeFunction. If x,y,z are floats returns a 2D array of size (N, 3). If given vectors, returns a 3D array of size (N, 3, length(x)).

Note

All coordinate vectors must be of equal length.

DDD.AbstractCrystalStruct — Type

abstract type AbstractCrystalStruct end
struct BCC <: AbstractCrystalStruct end
struct FCC <: AbstractCrystalStruct end
struct HCP <: AbstractCrystalStruct end

Crystal structures.

DDD.AbstractDlnSeg — Type

There are different types of segments. Edge segments are orthogonal to the Burgers vector; screw segments are parallel to the Burgers vector; mixed segments are anything in between. Segment type can be easily inferred during runtime. These are mainly used for multiple dispatch purposes.

abstract type AbstractDlnSeg end
struct segNone <: AbstractDlnSeg end    # Undefined segment.
struct segEdge <: AbstractDlnSeg end    # Edge segment.
struct segEdgeN <: AbstractDlnSeg end   # Edge segment parallel to slip plane.
struct segScrew <: AbstractDlnSeg end   # Screw segment.
struct segMixed <: AbstractDlnSeg end   # Mixed segment.

DDD.AbstractIntegrator — Type

abstract type AbstractIntegrator end
struct CustomTrapezoid <:AbstractIntegrator end

Integrator types.

DDD.AbstractMesh — Type

abstract type AbstractMesh end

Abstract mesh type.

DDD.AbstractMobility — Type

abstract type AbstractMobility end
struct BCC <: AbstractMobility end
struct FCC <: AbstractMobility end
struct HCP <: AbstractMobility end

Mobility functions.

DDD.AbstractShapeFunction — Type

abstract type AbstractShapeFunction end
abstract type AbstractShapeFunction3D <: AbstractShapeFunction end
abstract type AbstractShapeFunction2D <: AbstractShapeFunction end
struct LinearQuadrangle3D <:AbstractShapeFunction3D end
struct LinearQuadrangle2D <:AbstractShapeFunction2D end

Abstract types for different shape functions.

DDD.DislocationP — Type

DislocationP{
    T1 <: Float64,
    T2 <: Int64,
    T3 <: Bool,
    T4 <: AbstractMobility,
}
    # Size.
    coreRad::T1     # Core radius.
    coreRadMag::T1  # Magnitude of core Radius.
    # Connectivity.
    minSegLen::T1       # Minimum line length.
    maxSegLen::T1       # Maximum line length.
    minArea::T1         # Minimum area for remeshing.
    maxArea::T1         # Maximum area for remeshing.
    maxConnect::T2      # Maximum number of connections to a node.
    remesh::T3          # Flag for remeshing.
    collision::T3       # Flag for collision handling.
    separation::T3      # Flag for separation handling.
    virtualRemesh::T3   # Flag for virtual remeshing.
    # Mobility.
    edgeDrag::T1    # Drag coefficient edge dislocation.
    screwDrag::T1   # Drag coefficient screw dislocation.
    climbDrag::T1   # Drag coefficient climb.
    lineDrag::T1    # Drag coefficient line.
    mobility::T4    # Mobility law.

Dislocation parameters structure. See AbstractMobility for more details.

DDD.IntegrationP — Type

IntegrationP{
    T1 <: Float64,
    T2 <: AbstractIntegrator,
    T3 <: Int64,
}
    dt::T1
    tmin::T1
    tmax::T1
    method::T2
    abstol::T1
    reltol::T1
    time::T1
    step::T3

This structure contains the integration parameters for the simulation.

DDD.MaterialP — Type

MaterialP{T1 <: Float64, T2 <: AbstractCrystalStruct}
    μ::T1
    μMag::T1
    ν::T1
    E::T1
    crystalStruct::T2

DDD.coordIdx — Method

coordIdx(network::DislocationNetwork,
    index::Union{Integer,AbstractArray{<:Integer, N}) where {N}

Get coordinates for the node(s) that with the index or vector of indices provided.

DDD.coordLbl — Method

Related functions: idxLabel

coordLbl(network::DislocationNetwork, label::Integer)

Get coordinates for the nodes with a given label (node type).

DDD.dataCond — Method

Related functions: idxCond

dataCond(network::DislocationNetwork, dataField::Symbol, val::Real;
    condition::Function = ==)

Get the data whose dataField, (:links, :bVec, :slipPlane, :coord, :label, :numNode, :numSeg), meets the condition(dataField, val). If dataField is multidimensional it will use CartesianIndex to find it.

dataCond(network::DislocationNetwork, dataField::Symbol, idxComp::Integer,
    val::Real; condition::Function = ==)

Get the data whose dataField, meets the condition(dataField[:,idxComp], val). It errors if the fieldname provided does not have a column idxComp.

dataCond(network::DislocationNetwork, dataField::Symbol, condField::Symbol,
    val::Real; condition::Function = ==)

Get the data from dataField that corresponds to the condField (from the same pool as dataField) that meets the condition(condField, val). If condField is multidimensional it will use CartesianIndex in its search. dataField and condField must have the same number of rows.

dataCond(network::DislocationNetwork, dataField::Symbol, condField::Symbol,
    idxComp::Integer, val::Real; condition::Function = ==)

Get the data from dataField that corresponds to the condField that meets the condition(condField[:, idxComp], val). dataField and condField must have the same number of rows.

DDD.hatStress — Method

Calculate the reaction from a dislocation.

DDD.idxCond — Method

Related functions: dataCond

idxCond(network::DislocationNetwork, fieldname::Symbol,
    args...; condition::Function)

Find index/indices whose fieldname meets the condition(fieldname, args...) where condition can be any function that uses fieldname and args to make a comparison.

idxCond(data::Union{
        AbstractArray{<:Real,N1},
        AbstractArray{<:nodeType, {N2}}
    },
    val::Real; condition::Function = ==) where {N1,N2}

Find index/indices of data that meet the condition(data, val).

idxCond(network::DislocationNetwork, fieldname::Symbol, val::Real;
    condition::Function = ==)

Find index/indices of node whose fieldname, (:links, :bVec, :slipPlane, :coord, :label, :numNode, :numSeg), meets condition(fieldname, val). If fieldName is multidimensional it will return CartesianIndex.

idxCond(network::DislocationNetwork, fieldname::Symbol, idxComp::Integer,
    val::Real; condition::Function = ==)

Find index/indices of node whose fieldname meets condition(fieldname[:, idxComp], val). It errors if the fieldname provided does not have a column idxComp.

DDD.idxLabel — Method

Related functions: coordLbl

idxLabel(network::DislocationNetwork, label::Integer; condition::Function = ==)

Find indices for dislocations whose label meets condition(x, label).

DDD.inclusiveComparison — Method

inclusiveComparison(data, args...)

Compare data to a tuple, return true if it is equal to any arg, false if it is not equal to any.

Examples

julia> inclusiveComparison("f", 1,4,5,"f")
true
julia> inclusiveComparison(23.246, 1.5, 4, 5, "f")
false

DDD.limits! — Method

limits!(
    lims::AbstractArray{<:Float64, N1},
    segLen::Float64,
    range::AbstractArray{<:Float64, N2},
    buffer::Float64,
) where {N1, N2}

Calculate the spatial limits a dislocation will occupy. This and loopDistribution are used in translatePoints to distribute dislocations in the simulation domain.

DDD.makeConnect! — Method

makeConnect!(network::DislocationNetwork)

Creates connectivity and linksConnect matrices in-place. connectivity contains the number of other other nodes each node is connected to, up to maxConnect other nodes. It also contains the segments in which it's involved. linksConnect is the running total of the number of other nodes each node is connected to, will probably be deleted in the future. This is called from makeNetwork and makeNetwork!.

DDD.makeInstanceDict — Method

makeInstanceDict(valType::DataType)

Make a dictionary of enumerated variable instances. Helps in translating JSON files.

DDD.makeLoop — Method

makeLoop(
    loopType::T1,
    numSides::T2,
    nodeSide::T2,
    numLoops::T2,
    segLen::T3,
    slipSystem::T2,
    _slipPlane::T4,
    _bVec::T4,
    label::T5,
    buffer::T6,
    range::T7,
    dist::T8,
) where {
    T1 <: AbstractDlnStr,
    T2 <: Int64,
    T3 <: Union{T where {T <: Float64}, AbstractArray{<:Float64, N} where {N}},
    T4 <: AbstractArray{<:Float64, N} where {N},
    T5 <: Vector{nodeType},
    T6 <: Float64,
    T7 <: AbstractArray{<:Float64, N} where {N},
    T8 <: AbstractDistribution,
}

Constructor function for DislocationLoop. See AbstractDlnStr, AbstractDlnSeg, nodeType, AbstractDistribution for further details.

DDD.makeSegment — Method

makeSegment(type::AbstractDlnSeg, slipPlane::Vector{T}, bVec::Vector{T})
    where {T<:Float64}

Make signle segment depending on the segment type, see AbstractDlnSeg.

DDD.pkForce — Method

Calculate the Peach-Koehler force on segments.

DDD.subTypeTree — Method

subTypeTree(t, level = 1, dict = Dict(); cutoff = 0)

Create subtype dictionary. . Adapted from https://github.com/JuliaLang/julia/issues/24741

DDD.translateEnum — Method

translateEnum(
    valType::DataType,
    dict::Dict{T1, T2},
    key::T3,
) where {T1, T2, T3}

Translates the string name of enumerated types to the actual Julia type.

DDD.translatePoints — Method

function translatePoints(
    coord::AbstractArray{<:Float64, N1},
    lims::AbstractArray{<:Float64, N1},
    disp::AbstractArray{<:Float64, N2},
) where {N1, N2}

Translate dislocation node coordinates coord inside the spatial bounds of lims (calculated in limits!) according to the displacement disp (calculated in loopDistribution). Used to distribute sources inside a domain in makeNetwork and makeNetwork!.

JSON.Writer.lower — Method

JSON.lower(
    t::T,
) where {
    T <: Union{
        AbstractCrystalStruct,
        AbstractMobility,
        AbstractIntegrator,
        AbstractDlnSeg,
        AbstractDlnStr,
        AbstractDistribution,
    },
} = string(t)

JSON.lower(t::nodeType) = Int(t)

Extensions to JSON.lower for custom types. Allows these variables to be serialised properly.