Constructors

DDD.SlipSystem — Method

SlipSystem(crystalStruct::T1, slipPlane::T2, bVec::T2) 
    where {T1 <: AbstractCrystalStruct,T2}

Constructor for SlipSystem. Checks for orthogonality of the burgers vector and slip plane. It assumes each column corresponds to a slip system.

DDD.SlipSystem — Method

SlipSystem(; crystalStruct::T1, slipPlane::T2, bVec::T2) 
    where {T1 <: AbstractCrystalStruct,T2}

Keyword constructor for SlipSystem.

DDD.DislocationParameters — Method

DislocationParameters(
    coreRad::T1,
    coreRadMag::T1,
    minSegLen::T1,
    maxSegLen::T1,
    minArea::T1,
    maxArea::T1,
    edgeDrag::T1,
    screwDrag::T1,
    climbDrag::T1,
    lineDrag::T1,
    maxConnect::T2,
    mobility::T3,
    remesh::T4 = true,
    collision::T4 = true,
    separation::T4 = true,
    virtualRemesh::T4 = true,
    parCPU::T4 = false,
    parGPU::T4 = false,
    slipStepCritLen::T1 = maxSegLen / 2,
    slipStepCritArea::T1 = 0.5 * (slipStepCritLen^2) * sin(2 * π / 360),
) where {T1,T2 <: Integer,T3 <: AbstractMobility,T4 <: Bool}

The constructor for DislocationParameters provides a few default values and calculates derived quantities.

DDD.DislocationParameters — Method

DislocationParameters(;
    coreRad::T1,
    coreRadMag::T1,
    minSegLen::T1,
    maxSegLen::T1,
    minArea::T1,
    maxArea::T1,
    edgeDrag::T1,
    screwDrag::T1,
    climbDrag::T1,
    lineDrag::T1,
    maxConnect::T2,
    mobility::T3,
    remesh::T4 = true,
    collision::T4 = true,
    separation::T4 = true,
    virtualRemesh::T4 = true,
    parCPU::T4 = false,
    parGPU::T4 = false,
    slipStepCritLen::T1 = maxSegLen / 2,
    slipStepCritArea::T1 = 0.5 * (slipStepCritLen^2) * sin(2 * π / 360),
) where {T1,T2 <: Integer,T3 <: AbstractMobility,T4 <: Bool}

Keyword constructor for DislocationParameters. Calls the positional constructor internally.

DDD.DislocationLoop — Method

DislocationLoop(
    loopType::T1,
    numSides::T2,
    nodeSide::T2,
    numLoops::T2,
    segLen,
    slipSystem::T2,
    _slipPlane,
    _bVec,
    label::T3,
    buffer,
    range,
    dist::T4,
) where {
    T1 <: AbstractDlnStr,
    T2 <: Integer,
    T3 <: AbstractVector{nodeType},
    T4 <: AbstractDistribution,
}

General constructor for DislocationLoop.

DDD.DislocationLoop — Method

DislocationLoop(
    loopType::T1,
    numSides::T2,
    nodeSide::T2,
    numLoops::T2,
    segLen,
    slipSystem::T2,
    _slipPlane::T3,
    _bVec::T3,
    label::T4,
    buffer,
    range,
    dist::T5,
) where {
    T1 <: loopPure,
    T2 <: Integer,
    T3 <: AbstractArray{T,N} where {T,N},
    T4 <: AbstractVector{nodeType},
    T5 <: AbstractDistribution,
}

Constructor for loopPure loops DislocationLoop.

DDD.DislocationLoop — Method

DislocationLoop(
    loopType::T1,
    numSides::T2,
    nodeSide::T2,
    numLoops::T2,
    segLen,
    slipSystem::T2,
    _slipPlane::T3,
    _bVec::T3,
    label::T4,
    buffer,
    range,
    dist::T5,
) where {
    T1 <: loopImpure,
    T2 <: Integer,
    T3 <: AbstractArray{T,N} where {T,N},
    T4 <: AbstractVector{nodeType},
    T5 <: AbstractDistribution,
}

A fallback DislocationLoop constructor for other as of yet unimplemented loopType.

DDD.DislocationLoop — Method

DislocationLoop(;
    loopType::T1,
    numSides::T2,
    nodeSide::T2,
    numLoops::T2,
    segLen::T3,
    slipSystem::T2,
    _slipPlane::T4,
    _bVec::T4,
    label::T5,
    buffer::T6,
    range::T7,
    dist::T8,
) where {
    T1 <: AbstractDlnStr,
    T2 <: Integer,
    T3 <: Union{T where {T},AbstractArray{T,N} where {T,N}},
    T4 <: AbstractArray{T,N} where {T,N},
    T5 <: AbstractVector{nodeType},
    T6,
    T7 <: AbstractArray{T,N} where {T,N},
    T8 <: AbstractDistribution,
}

Keyword DislocationLoop constructor calls the positional constructor, which dispatches the appropriate method.

DDD.DislocationNetwork — Method

DislocationNetwork(
    links::T1,
    slipPlane::T2,
    bVec::T2,
    coord::T2,
    label::T3,
    nodeVel::T2,
    nodeForce::T2,
    numNode::T4 = zeros(Int, 1),
    numSeg::T4 = zeros(Int, 1),
    maxConnect::T5 = 4,
    connectivity::T1 = zeros(Int, 1 + 2 * maxConnect, length(label)),
    linksConnect::T1 = zeros(Int, 2, size(links, 2)),
    segIdx::T1 = zeros(Int, size(links, 2), 3),
    segForce::T6 = zeros(3, size(links)...),
) where {
    T1 <: AbstractArray{T,N} where {T,N},
    T2 <: AbstractArray{T,N} where {T,N},
    T3 <: AbstractVector{nodeType},
    T4 <: Union{Int,AbstractVector{Int}},
    T5 <: Integer,T6 <: AbstractArray{T,N} where {T,N},
}

DislocationNetwork constructor provides default values, validates inputs and calculates derived quantities.

DDD.DislocationNetwork — Method

DislocationNetwork(;
    links::T1,
    slipPlane::T2,
    bVec::T2,
    coord::T2,
    label::T3,
    nodeVel::T2,
    nodeForce::T2,
    numNode::T4 = zeros(Int, 1),
    numSeg::T4 = zeros(Int, 1),
    maxConnect::T5 = 0,
    connectivity::T1 = zeros(Int, 1 + 2 * maxConnect, length(label)),
    linksConnect::T1 = zeros(Int, 2, size(links, 2)),
    segIdx::T1 = zeros(Int, size(links, 2), 3),
    segForce::T6 = zeros(3, size(links, 2), 0),
) where {
    T1 <: AbstractArray{T,N} where {T,N},
    T2 <: AbstractArray{T,N} where {T,N},
    T3 <: AbstractVector{nodeType},T4 <: AbstractVector{Int},
    T5 <: Integer,
    T6 <: AbstractArray{T,N} where {T,N},
}

Keyword constructor for DislocationNetwork, calls the positional one.

DDD.DislocationNetwork — Method

DislocationNetwork(
    sources::T1,
    maxConnect::T2 = 4,
    args...;
    memBuffer = nothing,
    checkConsistency::T3 = true,
    kw...,
) where {
    T1 <: Union{T,
                AbstractVector{T},
                NTuple{N,T} where N
                } where {T <: DislocationLoop},
    T2 <: Integer,
    T3 <: Bool,
}

The recommended way of creating a network is to use a source of type DislocationLoop. It also accepts arrays and tuples of DislocationLoop variables. This automatically generates the network according to the parameters stored in source or each of its entries.

Meaning

args... are optional arguments that will be passed on to the loopDistribution function which distributes the loops in sources according to the type of their dist variable.

kw... are optional keyword arguments that will also be passed to loopDistribution.

memBuffer is the numerical value for allocating memory in advance. The quantity, memBuffer × N, where N is the total number of nodes in sources, will be the initial number of entries allocated in the matrices that keep the network's data. If no memBuffer is provided, the number of entries allocated will be `round(N*log2(N)).

DDD.DislocationNetwork! — Method

DislocationNetwork!(
    network::T1,
    sources::T2,
    maxConnect::T3 = 4,
    args...;
    checkConsistency::T4 = true,
    kw...,
) where {
    T1 <: DislocationNetwork,
    T2 <: Union{T,
                AbstractVector{T},
                NTuple{N,T} where N
                } where {T <: DislocationLoop},
    T3 <: Integer,
    T4 <: Bool,
}

Adds more sources to an existing network.

DDD.makeNetwork! — Function

makeNetwork!(
    links,
    slipPlane,
    bVec,
    coord,
    label,
    sources,
    lims,
    initIdx,
    args...;
    kw...,
)

Internal function called by the recommended DislocationNetwork method and its mutating counterpart DislocationNetwork!, to automatically fill the matrices with the information from sources.

Utility

DDD.loopDistribution — Function

loopDistribution(::Rand, n, args...; kw...)

Returns a 3 × n matrix whose points follow a random uniform distribution. Used by DislocationNetwork when the dist parameter of DislocationLoop is Rand().

loopDistribution(::Zeros, n, args...; kw...)

Returns a 3 × n zeros matrix. Used by DislocationNetwork when the dist parameter of DislocationLoop is Zeros().

loopDistribution(::Rand, n, args...; kw...)

Returns a 3 × n matrix whose points follow a random normal distribution. Used by DislocationNetwork when the dist parameter of DislocationLoop is Randn().

loopDistribution(::Rand, n, args...; kw...)

Returns a 3 × n matrix whose points follow are regularly placed. Used by DislocationNetwork when the dist parameter of DislocationLoop is Regular().

Note

Not yet implemented.

DDD.limits! — Function

limits!(lims, segLen, range, buffer)

Calculate the bounding limits within which a DislocationLoop can be distributed in a DislocationNetwork.

DDD.translatePoints! — Function

translatePoints!(coord, lims, disp)

Translates coordinates using the limits and displacements calculated by limits! and loopDistribution.

DDD.makeSegment — Function

makeSegment(::segEdge, slipPlane, bVec)

Return the edge segment (slipPlane × bVec) / || slipPlane × bVec ||. Used to create segments for DislocationNetwork.

makeSegment(::segEdge, slipPlane, bVec)

Return the edge segment slipPlane / || slipPlane ||. Used to create segments for DislocationNetwork.

makeSegment(::segScrew, slipPlane, bVec)

Return the screw segment bVec / || bVec ||. Used to create segments for DislocationNetwork.

DDD.makeConnect — Function

makeConnect(links, maxConnect)

Creates connectivity and linksConnect matrices for DislocationNetwork.

connectivity contains the number of other other nodes each node is connected to, up to maxConnect other nodes. Every (2i, j) entry contains a node connected to node j. Every (2i+1, j) coordinate contains whether that node is the first or second node in the link.

linksConnect relates connections enabled by a link. Analogous to the connectivity of a link.

DDD.makeConnect! — Function

makeConnect!(network::DislocationNetwork)

Mutates the connectivity and linksConnect matrices of network. Works the same as makeConnect.

DDD.getSegmentIdx — Function

getSegmentIdx(links, label)

Returns an n × 3 matrix is of the form [i, node1, node2].

i can be used to find the Burgers vector, slip plane and segment forces of segment i, eg bVec[:, i].

node1 and node2 can be used to find the coordinate and velocity of the nodes, eg l = coord[:, node2] - coord[:, node1].

DDD.getSegmentIdx! — Function

getSegmentIdx!(network::DislocationNetwork)

Mutates the segIdx matrix in network. Works the same way as getSegmentIdx.

DDD.checkNetwork — Function

checkNetwork(network::DislocationNetwork)

Checks the validity of the dislocation network. It ensures the following conditions are met by the member variables of network:

connectivity and links have the same number of non-zero entries;

all entries in bVec are non-zero;

only the trailing columns of connectivity are zeros;

consistency between connectivity and links;

bVec is conserved among connected nodes;

entries in links are unique;

consistency betwen connectivity and linksConnect