Dislocation Generation

Types, Structs and Constructors

Dislocations are described, generated and validated by custom types, structures and functions. By subtyping the provided types with new concrete types, users can define functions which dispatch specifically on their new types while minimising the need for code rewrites, as multiple dispatch takes care of everything during JIT compilation. Structures have not had their default constructors overwritten, we provide custom constructors whose use is recommended instead.

In discrete dislocation dynamics, dislocations are described by nodes connected by segments. The nodes are labelled according to their type, which is used by the software to decide how they are treated. However, labels are discrete variables, so they cannot take on any value. Additionally, accidentally using non-existent node types may produce silent and difficult to track errors. It is also impractical to validate node types at runtime. We solve these issues by defining a custom enumerated type, which not only limits possible values but informs users and developers of what the values represent.

DDD.nodeType — Type

@enum nodeType begin
    none = 0    # Undefined node, value at initialisation.
    intMob = 1  # Internal mobile node.
    intFix = 2  # Internal fixed node.
    srfMob = 3  # Mobile surface node.
    srfFix = 4  # Fixed surface node.
    ext = 5     # External node.
end

Of course, mislabelling a node with an erroneous but defined value may still occur. Preventing such bugs is the task of users and developers, however the problem may is eased by the self-descriptive nature of enumerated types.

Dislocations also have different idealised segment types which are characterised by the relationship between the line direction and Burgers vector. These idealised types are used in the code for in loop generation. We've defined a few common types, most of which are not currently used but may prove useful in the future, for example in the statistical analysis of the dislocation network.

DDD.AbstractDlnSeg — Type

abstract type AbstractDlnSeg end
struct segNone <: AbstractDlnSeg end    # Undefined segment
struct segEdge <: AbstractDlnSeg end    # Edge segment
struct segEdgeN <: AbstractDlnSeg end   # Edge segment
struct segScrew <: AbstractDlnSeg end   # Screw segment
struct segMixed <: AbstractDlnSeg end   # Mixed segment

where segEdge have $(\bm{b} \perp \bm{t}) \perp \bm{n}$, segEdgeN have $(\bm{b} \perp \bm{t}) \parallel \bm{n}$, segScrew have $\bm{b} \parallel \bm{t}$, segMixed have $\bm{b} \not\perp \bm{t}~ \&~ \bm{b} \not\perp \bm{n}$ and $\bm{b}$ is the Burgers vector and $\bm{n}$ the slip plane.

Dislocation loops are idealised as having different classifications. Prismatic loops are made up only of edge segments generally with the same slip system; shear loops are made up of a mixture of segment types with the same slip system; jogs and kinks are steps not contained in the slip plane. These idealisations can be used to automate loop generation with minimal rewriting via multiple dispatch. We provide the following types for such a purpose.

DDD.AbstractDlnStr — Type

abstract type AbstractDlnStr end
struct loopDln <: AbstractDlnStr end    # Unclassified loop
struct loopPrism <: AbstractDlnStr end  # Prismatic loop
struct loopShear <: AbstractDlnStr end  # Shear loop
struct loopJog <: AbstractDlnStr end    # Jog
struct loopKink <: AbstractDlnStr end   # Kink

Generating dislocation networks often involves distributing the initial loops within the simulation domain in a particular way. We again define custom structures that enable us to make use of multiple dispatch.

DDD.AbstractDistribution — Type

abstract type AbstractDistribution end
struct Zeros <: AbstractDistribution end
struct Rand <: AbstractDistribution end
struct Randn <: AbstractDistribution end
struct Regular <: AbstractDistribution end

Distributions for dislocation sources.

Mobility functions describe how dislocations move within a material. There are many variations of such functions and users may want/need to use different functions for different purposes. Creating concrete mobility types lets users define a function for their new concrete mobility type and carry on with their lives.

DDD.AbstractMobility — Type

abstract type AbstractMobility end
struct mobBCC <: AbstractMobility end
struct mobFCC <: AbstractMobility end
struct mobHCP <: AbstractMobility end

Mobility functions.

As previously mentioned, idealised dislocation segments live on slip systems, which are pairings of slip plane and Burgers vector. These can be stored in the following structure.

DDD.SlipSystem — Type

struct SlipSystem{T1, T2}
    crystalStruct::T1   # Crystal structure
    slipPlane::T2       # Slip plane
    bVec::T2            # Burgers vector
end

Structure to store slip systems.

As slip systems are defined with respect to pure edge dislocations, we recommended users use the keyword constructor as it validates the orthogonality of paired Burgers vector and slip plane.

DDD.SlipSystem — Method

SlipSystem(;
    crystalStruct::T1,
    slipPlane::T2,
    bVec::T2,
) where {T1 <: AbstractCrystalStruct, T2 <: AbstractArray{T, N} where {T, N}}

Keyword constructor for SlipSystem. Throws error if $\bm{b} \not\perp \bm{n}$ where $\bm{b}$ is the Burgers vector and $\bm{n}$ the slip plane.

The simulation requires certain parameters pertaining to the dislocation network being modelled. These values control certain aspects of the simulation and are stored in the following structure.

DDD.DislocationP — Type

struct DislocationP{T1, T2, T3, T4}
    coreRad::T1         # Core radius
    coreRadSq::T1       # Square of core radius
    coreRadMag::T1      # Magnitude of core radius
    minSegLen::T1       # Minimum segment length
    maxSegLen::T1       # Maximum segment length
    twoMinSegLen::T1    # Twice minimum segment length
    minArea::T1         # Minimum area enclosed by 3 segments
    maxArea::T1         # Maximum area enclosed by 3 segments
    minAreaSq::T1       # Squared min area
    maxAreaSq::T1       # Squared max area
    maxConnect::T2      # Maximum connectivity
    remesh::T3          # Remesh flag
    collision::T3       # Collision flag
    separation::T3      # Separation flag
    virtualRemesh::T3   # Virtual remeshing flag
    edgeDrag::T1        # Drag coefficient edge dislocation
    screwDrag::T1       # Drag coefficient screw dislocation
    climbDrag::T1       # Drag coefficient climb direction
    lineDrag::T1        # Drag coefficient line direction
    mobility::T4        # Mobility law
end

Structure to store dislocation parameters.

We recommend the use of the keyword constructor as it performs sanity checks on various parameters, ensuring a hierarchy of values is maintained.

DDD.DislocationP — Method

DislocationP(;
    coreRad::T1,
    coreRadMag::T1,
    minSegLen::T1,
    maxSegLen::T1,
    minArea::T1,
    maxArea::T1,
    maxConnect::T2,
    remesh::T3,
    collision::T3,
    separation::T3,
    virtualRemesh::T3,
    edgeDrag::T1,
    screwDrag::T1,
    climbDrag::T1,
    lineDrag::T1,
    mobility::T4,
) where {T1, T2 <: Int, T3 <: Bool, T4 <: AbstractMobility}

Keyword constructor for DislocationP. Validates values and calculates derived quantities.

Dislocation loops form the basis of a network, we provide a structure to store these loops, whether idealised or otherwise.

DDD.DislocationLoop — Type

struct DislocationLoop{T1, T2, T3, T4, T5, T6, T7, T8, T9}
    loopType::T1    # Loop type
    numSides::T2    # Number of sides per loop
    nodeSide::T2    # Number of nodes per side
    numLoops::T2    # Number of loops to generate
    segLen::T3      # Segment lengths
    slipSystem::T4  # Slip system
    links::T5       # Links matrix
    slipPlane::T6   # Slip plane for each link
    bVec::T6        # Burgers vector for each link
    coord::T6       # Coordinates of each node
    label::T7       # Label of each node
    buffer::T8      # Mean distance buffer separating each loop centre
    range::T6       # Distribution range of generated loops
    dist::T9        # Distribution of generated loops
end

Structure to store dislocation loop.

Again we recommend the use of the key-word only constructor as it provides an interface to call specific constructors which dispatch on loopType and generate the loop automatically. The default constructor should only be used for truly custom loops, however creating a concrete subtype of AbstractDlnStr and a constructor which dispatches on this new type is highly recommended since it facilitates testing, reproducibility and seamlessly integrates with existing infrastructure.

DDD.DislocationLoop — Method

DislocationLoop(;
    loopType::T1,
    numSides::T2,
    nodeSide::T2,
    numLoops::T2,
    segLen::T3,
    slipSystem::T2,
    _slipPlane::T4,
    _bVec::T4,
    label::T5,
    buffer::T6,
    range::T7,
    dist::T8,
) where {
    T1 <: AbstractDlnStr,
    T2 <: Int,
    T3 <: Union{T where {T}, AbstractArray{T, N} where {T, N}},
    T4 <: AbstractArray{T, N} where {T, N},
    T5 <: AbstractVector{nodeType},
    T6,
    T7 <: AbstractArray{T, N} where {T, N},
    T8 <: AbstractDistribution,
}

Generic keyword constructor for DislocationLoop. Calls other constructors that dispatch on loopType.

The first concrete DislocationLoop constructor is the "zero" constructor. The way Julia's multiple dispatch works is by dispatching on the most specific method for the inputs provided. Therefore this constructor will be called whenever the loop type is loopDln().

DDD.DislocationLoop — Method

DislocationLoop(;
    loopType::T1,
    numSides::T2,
    nodeSide::T2,
    numLoops::T2,
    segLen::T3,
    slipSystem::T2,
    _slipPlane::T4,
    _bVec::T4,
    label::T5,
    buffer::T6,
    range::T7,
    dist::T8,
) where {
    T1 <: AbstractDlnStr,
    T2 <: Int,
    T3 <: Union{T where {T}, AbstractArray{T, N} where {T, N}},
    T4 <: AbstractArray{T, N} where {T, N},
    T5 <: AbstractVector{nodeType},
    T6,
    T7 <: AbstractArray{T, N} where {T, N},
    T8 <: AbstractDistribution,
}

Generic keyword constructor for DislocationLoop. Calls other constructors that dispatch on loopType.

DislocationLoop(
    loopType::T1;
    numSides::T2,
    nodeSide::T2,
    numLoops::T2,
    segLen::T3,
    slipSystem::T2,
    _slipPlane::T4,
    _bVec::T4,
    label::T5,
    buffer::T6,
    range::T7,
    dist::T8,
) where {
    T1 <: loopDln,
    T2 <: Int,
    T3 <: Union{T where {T}, AbstractArray{T, N} where {T, N}},
    T4 <: AbstractArray{T, N} where {T, N},
    T5 <: AbstractVector{nodeType},
    T6,
    T7 <: AbstractArray{T, N} where {T, N},
    T8 <: AbstractDistribution,
}

Constructor for a "zero" DislocationLoop.

DislocationLoop(
    loopType::T1;
    numSides::T2,
    nodeSide::T2,
    numLoops::T2,
    segLen::T3,
    slipSystem::T2,
    _slipPlane::T4,
    _bVec::T4,
    label::T5,
    buffer::T6,
    range::T7,
    dist::T8,
) where {
    T1 <: AbstractDlnStr,
    T2 <: Int,
    T3 <: Union{T where {T}, AbstractArray{T, N} where {T, N}},
    T4 <: AbstractArray{T, N} where {T, N},
    T5 <: AbstractVector{nodeType},
    T6,
    T7 <: AbstractArray{T, N} where {T, N},
    T8 <: AbstractDistribution,
}

Validates inputs and generates a DislocationLoop of loopType defined by the arguments.

We also provide a catch-all constructor that generates shear or prismatic loops depending on whether loopType is loopShear() or loopPrism(). It acts as a fallback for other loop types but generates prismatic loops. Such behaviour can be overridden by defining new methods which dispatch on more a specific loopType.

DDD.DislocationLoop — Method

DislocationLoop(;
    loopType::T1,
    numSides::T2,
    nodeSide::T2,
    numLoops::T2,
    segLen::T3,
    slipSystem::T2,
    _slipPlane::T4,
    _bVec::T4,
    label::T5,
    buffer::T6,
    range::T7,
    dist::T8,
) where {
    T1 <: AbstractDlnStr,
    T2 <: Int,
    T3 <: Union{T where {T}, AbstractArray{T, N} where {T, N}},
    T4 <: AbstractArray{T, N} where {T, N},
    T5 <: AbstractVector{nodeType},
    T6,
    T7 <: AbstractArray{T, N} where {T, N},
    T8 <: AbstractDistribution,
}

Generic keyword constructor for DislocationLoop. Calls other constructors that dispatch on loopType.

DislocationLoop(
    loopType::T1;
    numSides::T2,
    nodeSide::T2,
    numLoops::T2,
    segLen::T3,
    slipSystem::T2,
    _slipPlane::T4,
    _bVec::T4,
    label::T5,
    buffer::T6,
    range::T7,
    dist::T8,
) where {
    T1 <: loopDln,
    T2 <: Int,
    T3 <: Union{T where {T}, AbstractArray{T, N} where {T, N}},
    T4 <: AbstractArray{T, N} where {T, N},
    T5 <: AbstractVector{nodeType},
    T6,
    T7 <: AbstractArray{T, N} where {T, N},
    T8 <: AbstractDistribution,
}

Constructor for a "zero" DislocationLoop.

DislocationLoop(
    loopType::T1;
    numSides::T2,
    nodeSide::T2,
    numLoops::T2,
    segLen::T3,
    slipSystem::T2,
    _slipPlane::T4,
    _bVec::T4,
    label::T5,
    buffer::T6,
    range::T7,
    dist::T8,
) where {
    T1 <: AbstractDlnStr,
    T2 <: Int,
    T3 <: Union{T where {T}, AbstractArray{T, N} where {T, N}},
    T4 <: AbstractArray{T, N} where {T, N},
    T5 <: AbstractVector{nodeType},
    T6,
    T7 <: AbstractArray{T, N} where {T, N},
    T8 <: AbstractDistribution,
}

Validates inputs and generates a DislocationLoop of loopType defined by the arguments.

The dist parameter refers to a concrete subtype of AbstractDistribution. When defining a new distribution it is important to define a new version of the loopDistribution function.

DDD.loopDistribution — Function

loopDistribution(dist<:AbstractDistribution, n::Int, args...; kw...)
loopDistribution(dist::Zeros, n::Int, args...; kw...) = zeros(3, n)
loopDistribution(dist::Rand, n::Int, args...; kw...) = rand(3, n)
loopDistribution(dist::Randn, n::Int, args...; kw...) = randn(3, n)
loopDistribution(dist::Regular, n::Int, args...; kw...) = error("loopDistribution: regular distribution yet not implemented")

Returns a 3 × n matrix whose points follow the distribution defined by the method dispatching on the concrete subtype of dist. When creating custom AbstractDistribution subtypes, a corresponding loopDistribution method must be created for the custom distribution to be used in the DislocationLoop constructors.

The dislocation loops contain all the data relevant to a single loop. This data is then used to populate a dislocation network, which is a mutable structure because it evolves over time.

DDD.DislocationNetwork — Type

mutable struct DislocationNetwork{T1, T2, T3, T4, T5, T6}
    links::T1
    slipPlane::T2
    bVec::T2
    coord::T2
    label::T3
    nodeVel::T2
    numNode::T4         # Number of nodes in network
    numSeg::T4          # Number of segments in network
    maxConnect::T4      # Maximum connections per node
    connectivity::T5    # Connectivity matrix
    linksConnect::T5    # Links involved in connection
    segIdx::T5          # Contains segment index and the nodes of the nodes in said link
    segForce::T6        # Force on each node of each segment
end

Structure to store dislocation network.

Again we provide a keyword constructor which performs some sanity checks and loads the data into the structure. This is the constructor to use when loading data from a previously generated network.

DDD.DislocationNetwork — Method

DislocationNetwork(;
    links::T1,
    slipPlane::T2,
    bVec::T2,
    coord::T2,
    label::T3,
    nodeVel::T2,
    numNode::T4 = 0,
    numSeg::T4 = 0,
    maxConnect::T4 = 0,
    connectivity::T5 = zeros(Int, 0, 0),
    linksConnect::T5 = zeros(Int, 2, 0),
    segIdx::T5 = zeros(Int, 2, 3),
    segForce::T6 = zeros(3, 2, 0),
) where {
    T1 <: AbstractArray{T, N} where {T, N},
    T2 <: AbstractArray{T, N} where {T, N},
    T3 <: AbstractVector{nodeType},
    T4 <: Int,
    T5 <: AbstractArray{Int, N} where {N},
    T6 <: AbstractArray{T, N} where {T, N},
}

Keyword constructor for DislocationNetwork, performs validations but creates dislocation network as provided.

However, if the aim is to generate a new network then use the following constructor.

DDD.DislocationNetwork — Method

DislocationNetwork(
    sources::T1,
    maxConnect::T2 = 4,
    args...;                        # Optional arguments
    memBuffer = nothing,            # Buffer for memory allocation
    checkConsistency::T3 = true,    # Check consistency of generated network
    kw...,                          # Other keyword arguments
) where {
    T1 <: Union{T, AbstractVector{T}} where {T <: DislocationLoop},
    T2 <: Int,
    T3 <: Bool,
}

Out of place constructor for DislocationNetwork. Generates a new dislocation network from previously generated sources.

Argument Explanation

args... are optional arguments that will be passed on to the loopDistribution function which distributes the loops in sources according to the type of their dist variable.

kw... are optional keyword arguments that will also be passed to loopDistribution.

memBuffer is the numerical value for allocating memory in advance, the quantity $\textrm{memBuffer} \times N$ where N is the total number of nodes in sources, will be the initial number of entries allocated in the matrices that keep the network's data, if it is nothing then the number of entries is $\textrm{round}(N \log_{2}(N))$.

If adding to an existing network, use the mutating (also called in-place) constructor.

DDD.DislocationNetwork! — Method

DislocationNetwork!(
    network::T1,
    sources::T2,
    maxConnect::T3 = 4,
    args...;
    checkConsistency::T4 = true,
    kw...,
) where {
    T1 <: DislocationNetwork,
    T2 <: Union{T, AbstractVector{T}} where {T <: DislocationLoop},
    T3 <: Int,
    T4 <: Bool,
}

In-place constructor for DislocationNetwork. Generates a new dislocation network from already generated sources. If the matrices already in network are not large enough to accommodate the additions from sources, it will automatically allocate $\textrm{round}(N \log_{2}(N))$ new entries where N is the total number of nodes in sources.

Dislocation network constructors use a few internal functions to distribute loops about the domain as well as create auxiliary matrices and verify the integrity of the generated network. As previously mentioned, loopDistribution is used to generate points from a particular distribution. These points must be scaled and adjusted by limits generated the limits function.

DDD.limits! — Method

limits!(
    lims::T1,
    segLen::T2,
    range::T1,
    buffer::T2,
) where {T1 <: AbstractArray{T, N} where {T, N}, T2}

In-place addition of buffer × segLen to range in order to calculate the limits in which dislocations will exist.

The limits, together with the aforementioned distributions are used to translate coordinates with the translatePoints function.

DDD.translatePoints — Method

translatePoints(
    coord::T1,
    lims::T1,
    disp::T2,
) where {T1 <: AbstractArray{T, N} where {T, N}, T2 <: AbstractVector{T} where {T}}

Translates coordinates using the limits and displacements calculated by limits! and loopDistribution.

In order to traverse the network, it is useful to define a few auxiliary matrices containing relational information about nodes and links. These are created by the makeConnect functions.

DDD.makeConnect — Method

makeConnect(
    links::T1,
    maxConnect::T2,
) where {T1 <: AbstractArray{T, N} where {T, N}, T2 <: Int}

Creates connectivity and linksConnect matrices. connectivity contains the number of other other nodes each node is connected to, up to maxConnect other nodes. Every (2i, j) entry contains a node connected to node j. Every (2i+1, j) coordinate contains whether that node is the first or second node in the link.

The matrix linksConnect relates connections enabled by a link. Analogous to the connectivity of a link.

DDD.makeConnect! — Method

makeConnect!(network::DislocationNetwork)

In-place version of makeConnect.

It's also useful to define another matrix for indexing segments quickly, this matrix is defined by the getSegmentIdx functions.

DDD.getSegmentIdx — Method

getSegmentIdx(
    links::T1,
    label::T2,
) where {T1 <: AbstractArray{T, N} where {T, N}, T2 <: AbstractVector{nodeType}}

Creates an indexing matrix for quick indexing of dislocation segments for quick access to slip planes, burgers vectors and line vectors. The return 3 × n matrix is of the form [i, node1, node2]. Index i can be used to find the Burgers vector, slip plane and segment forces of a segment, eg bVec[:, i]. While node1 and node2 can be used to find the coordinate and velocity of the nodes, eg t = coord[:, node2] - coord[:, node1].

DDD.getSegmentIdx! — Method

getSegmentIdx!(network::DislocationNetwork)

In-place version of getSegmentIdx.

The validity of the network can be checked by checkNetwork.

DDD.checkNetwork — Method

checkNetwork(network::DislocationNetwork)

Checks the validity of the dislocation network. It ensures the following conditions are met by the member variables of network:

connectivity and links have the same number of non-zero entries;

all entries in bVec are non-zero;

only the trailing columns of connectivity are zeros;

consistency between connectivity and links;

bVec is conserved among connected nodes;

entries in links are unique;

consistency betwen connectivity and linksConnect