Constructors

DDD.FEMParameters — Method

FEMParameters(
    type::T1,
    order::T2,
    dx::T3,
    dy::T3,
    dz::T3,
    mx::T4,
    my::T4,
    mz::T4
) where {
    T1 <: AbstractMesh,
    T2 <: AbstractElementOrder,
    T3 <: AbstractFloat,
    T4 <: Integer
}

Constructor for FEMParameters.

DDD.FEMParameters — Method

FEMParameters(;
    type::T1,
    order::T2,
    dx::T3,
    dy::T3,
    dz::T3,
    mx::T4,
    my::T4,
    mz::T4
) where {
    T1 <: AbstractMesh,
    T2 <: AbstractElementOrder,
    T3 <: AbstractFloat,
    T4 <: Integer
}

Keyword constructor for FEMParameters. Calls the positional constructor.

DDD.buildMesh — Function

buildMesh(matParams::T1, femParams::T2) 
    where {T1 <: MaterialParameters,T2 <: FEMParameters}

Internally calls another buildMesh that dispatches based on FEMParameters.type.

buildMesh(::T1, matParams::T2, femParams::T3) 
    where {T1 <: DispatchRegularCuboidMesh,T2 <: MaterialParameters,T3 <: FEMParameters}

Builds a RegularCuboidMesh by dispatching on femParams.type == DispatchRegularCuboidMesh().

DDD.RegularCuboidMesh — Method

RegularCuboidMesh(
    order::T1, 
    matParams::T2, 
    femParams::T3
) where {
    T1 <: LinearElement,
    T2 <: MaterialParameters,
    T3 <: FEMParameters
}

Created by: E. Tarleton edmund.tarleton@materials.ox.ac.uk

3D FEM code using linear 8 node element with 8 integration pts (2x2x2) per element.

   4.-------.3
   | \       |\
   |  \      | \    my
   1.--\---- .2 \
    \   \     \  \
     \  8.--------.7
      \  |      \ |  mz
       \ |       \|
        5.--------.6
             mx
y   ^z
 ↖  |
  \ |
   \|---->x

Note

This is rotated about the x axis w.r.t. to the local (s1, s2, s3) system. calc_σHat uses custom linear shape functions that eliminate the need for a Jacobian, speeding up the calculation. We keep the Jacobian in this function because it's only run at simulation initialisation so it's not performance critical, which lets us use standard shape functions.

DDD.RegularCuboidMesh — Method

RegularCuboidMesh(;
    order::T1, 
    matParams::T2, 
    femParams::T3
) where {
    T1 <: AbstractElementOrder,
    T2 <: MaterialParameters,
    T3 <: FEMParameters
}

Keyword constructor for RegularCuboidMesh.