Forces

DDD.calcSegForce — Function

calcSegForce(
    dlnParams::DislocationParameters,
    matParams::MaterialParameters,
    network::DislocationNetwork;
    # mesh::RegularCuboidMesh,
    # dlnFEM::DislocationFEMCorrective;
    parallel::Bool = true,
)

Missing docstring.

Missing docstring for calcSegForce!. Check Documenter's build log for details.

DDD.calcPKForce — Function

calcPKForce(
    mesh::RegularCuboidMesh,
    dlnFEM::DislocationFEMCorrective,
    network::DislocationNetwork,
)

Calculate the Peach-Koehler force on segments.

$f = (\hat{\mathbb{\sigma}} \cdot \overrightarrow{b}) \times \overrightarrow{t}$

Missing docstring.

Missing docstring for calcPKForce!. Check Documenter's build log for details.

DDD.calc_σHat — Function

calc_σHat(
    mesh::RegularCuboidMesh,
    forceDisplacement::ForceDisplacement,
    x0::AbstractVector{T} where {T},
)

Calculate the reaction stress, $̂σ$, from a dislocation.

DDD.calcSelfForce — Function

calcSelfForce(
    dlnParams::DislocationParameters,
    matParams::MaterialParameters,
    network::DislocationNetwork,
)

Calculates the self-interaction force felt by two nodes in a segment. Naturally the forces are equal and opposite to each other.

Missing docstring.

Missing docstring for calcSelfForce!. Check Documenter's build log for details.

DDD.calcSegSegForce — Function

!!! Note This function is based on the SegSegForces function by A. Arsenlis et al. It is optimised for speed and reusability. It has also been locally parallelised.

It implements the analytical solution of the force between two dislocation segments. Details are found in Appendix A.1. in "Enabling Strain Hardening Simulations with Dislocation Dynamics" by A. Arsenlis et al.

At a high level this works by creating a local coordinate frame using the line directions of the dislocation segments and a vector orthogonal to them. The line integrals are then evaluated parametrically utilising this local coordinate. BibTex citation here:

@article{Arsenlis_2007, doi = {10.1088/0965-0393/15/6/001}, url = {https://doi.org/10.1088%2F0965-0393%2F15%2F6%2F001}, year = 2007, month = {jul}, publisher = {{IOP} Publishing}, volume = {15}, number = {6}, pages = {553–595}, author = {A Arsenlis and W Cai and M Tang and M Rhee and T Oppelstrup and G Hommes and T G Pierce and V V Bulatov}, title = {Enabling strain hardening simulations with dislocation dynamics}, journal = {Modelling and Simulation in Materials Science and Engineering}, abstract = {Numerical algorithms for discrete dislocation dynamics simulations are investigated for the purpose of enabling strain hardening simulations of single crystals on massively parallel computers. The algorithms investigated include the calculation of forces, the equations of motion, time integration, adaptive mesh refinement, the treatment of dislocation core reactions and the dynamic distribution of data and work on parallel computers. A simulation integrating all these algorithmic elements using the Parallel Dislocation Simulator (ParaDiS) code is performed to understand their behaviour in concert and to evaluate the overall numerical performance of dislocation dynamics simulations and their ability to accumulate percent of plastic strain.} }

Missing docstring.

Missing docstring for calcSegSegForce!. Check Documenter's build log for details.