DDD.jl

DDD.AbstractCrystalStruct

DDD.AbstractDistribution

DDD.AbstractDlnSeg

DDD.AbstractDlnStr

DDD.AbstractIntegrator

DDD.AbstractMobility

DDD.DislocationLoop

DDD.DislocationNetwork

DDD.DislocationP

DDD.nodeType

Base.isequal

DDD.checkNetwork

DDD.compStruct

DDD.coordIdx

DDD.coordLbl

DDD.dataCond

DDD.extAngle

DDD.idxCond

DDD.idxLabel

DDD.inclusiveComparison

DDD.intAngle

DDD.limits!

DDD.loopDistribution

DDD.makeConnect

DDD.makeLoop

DDD.makeNetwork

DDD.makeNetwork!

DDD.makeSegment

DDD.pushToDataFrame!

DDD.rot3D

DDD.saveParams

DDD.translatePoints

DDD.@makeType

DDD.@string_as_varname_macro

DDD.AbstractCrystalStruct — Type

abstract type AbstractCrystalStruct end
struct BCC <:AbstractCrystalStruct end
struct FCC <:AbstractCrystalStruct end
struct HCP <:AbstractCrystalStruct end

Crystal structures.

DDD.AbstractDistribution — Type

abstract type AbstractDistribution end
struct Zeros <: AbstractDistribution end
struct Rand <: AbstractDistribution end
struct Randn <: AbstractDistribution end
struct Regular <: AbstractDistribution end

Distributions for dislocation sources.

DDD.AbstractDlnSeg — Type

abstract type AbstractDlnSeg end
struct segNone <: AbstractDlnSeg end
struct segEdge <: AbstractDlnSeg end
struct segEdgeN <: AbstractDlnSeg end
struct segScrew <: AbstractDlnSeg end
struct segMixed <: AbstractDlnSeg end

Dislocation segment types.

DDD.AbstractDlnStr — Type

abstract type AbstractDlnStr end
struct loopPrism <: AbstractDlnStr end
struct loopShear <: AbstractDlnStr end
struct loopDln <: AbstractDlnStr end
struct loopJog <: AbstractDlnStr end
struct loopKink <: AbstractDlnStr end

Idealised dislocation structure types.

DDD.AbstractIntegrator — Type

abstract type AbstractIntegrator end
struct CustomTrapezoid <:AbstractIntegrator end

Integrator types.

DDD.AbstractMobility — Type

abstract type AbstractMobility end
struct BCC <:AbstractMobility end
struct FCC <:AbstractMobility end
struct HCP <:AbstractMobility end

Mobility functions.

DDD.DislocationLoop — Type

DislocationLoop{
    T1 <: AbstractDlnStr,
    T2 <: Int64,
    T3 <: Union{
        T where {T <: AbstractDlnSeg},
        AbstractArray{<:AbstractDlnSeg, N} where {N},
    },
    T4 <: Union{
        T where {T <: Float64},
        AbstractArray{<:Float64, N} where {N}
    },
    T5 <: Union{Int64, AbstractArray{<:Int64, N} where {N}},
    T6 <: AbstractArray{<:Int64, N} where {N},
    T7 <: AbstractArray{<:Float64, N} where {N},
    T8 <: Vector{<:nodeType},
    T9 <: Float64,
    T10 <: AbstractDistribution,
}
    loopType::T1
    numSides::T2
    nodeSide::T2
    numLoops::T2
    segType::T3
    segLen::T4
    slipSystem::T5  # Slip system/systems of segments.
    links::T6       # Link matrix for dislocation nodes.
    slipPlane::T7   # Slip planes of all segments in loop.
    bVec::T7        # Burgers vector of all segments in loop.
    coord::T7       # Coords of all nodes in loop.
    label::T8
    buffer::T9
    range::T7
    dist::T10

Dislocation loop structure generated via the constructor makeLoop.

DDD.DislocationNetwork — Type

DislocationNetwork{
    T1 <: AbstractArray{<:Int64, N} where {N},
    T2 <: AbstractArray{<:Float64, N} where {N},
    T3 <: Vector{nodeType},
    T4 <: Int64,
    T5 <: Integer,
}
    links::T1
    slipPlane::T2
    bVec::T2
    coord::T2
    label::T3
    numNode::T4 = 0     # Total number of nodes in network.
    numSeg::T4 = 0      # Total number of segs in network.
    maxConnect::T5 = 4  # Maximum connectivity of nodes.
    connectivity::T1
    linksConnect::T1

Dislocation Network structure. See DislocationLoop, makeNetwork and makeNetwork! for further details.

DDD.DislocationP — Type

DislocationP{
    T1 <: Float64,
    T2 <: Int64,
    T3 <: Bool,
    T4 <: AbstractMobility,
}
    # Size.
    coreRad::T1     # Core radius.
    coreRadMag::T1  # Magnitude of core Radius.
    # Connectivity.
    minSegLen::T1       # Minimum line length.
    maxSegLen::T1       # Maximum line length.
    minArea::T1         # Minimum area for remeshing.
    maxArea::T1         # Maximum area for remeshing.
    maxConnect::T2      # Maximum number of connections to a node.
    remesh::T3          # Flag for remeshing.
    collision::T3       # Flag for collision handling.
    separation::T3      # Flag for separation handling.
    virtualRemesh::T3   # Flag for virtual remeshing.
    # Mobility.
    edgeDrag::T1    # Drag coefficient edge dislocation.
    screwDrag::T1   # Drag coefficient screw dislocation.
    climbDrag::T1   # Drag coefficient climb.
    lineDrag::T1    # Drag coefficient line.
    mobility::T4    # Mobility law.

Dislocation parameters structure. See AbstractMobility for more details.

DDD.nodeType — Type

@enum nodeType begin
    undef = -1
    intMob = 0
    intFix = 1
    srfMob = 2
    srfFix = 3
    ext = 4
end

Enumerated type for dislocation nodes.

DDD.checkNetwork — Method

checkNetwork(network::DislocationNetwork)

Checks the validity of the dislocation network. It ensures the following conditions are met by the member variables of network:

connectivity and links have the same number of non-zero entries;

all entries in bVec are non-zero;

only the trailing columns of connectivity are zeros;

consistency between connectivity and links;

bVec is conserved among connected nodes;

entries in links are unique;

consistency betwen connectivity and linksConnect

DDD.compStruct — Method

compStruct(arg1, arg2)

Compares structures to see if they are equal.

DDD.coordIdx — Method

coordIdx(network::DislocationNetwork,
    index::Union{Integer,AbstractArray{<:Integer, N}) where {N}

Get coordinates for the node(s) that with the index or vector of indices provided.

DDD.coordLbl — Method

Related functions: idxLabel

coordLbl(network::DislocationNetwork, label::Integer)

Get coordinates for the nodes with a given label (node type).

DDD.dataCond — Method

Related functions: idxCond

dataCond(network::DislocationNetwork, dataField::Symbol, val::Real;
    condition::Function = ==)

Get the data whose dataField, (:links, :bVec, :slipPlane, :coord, :label, :numNode, :numSeg), meets the condition(dataField, val). If dataField is multidimensional it will use CartesianIndex to find it.

dataCond(network::DislocationNetwork, dataField::Symbol, idxComp::Integer,
    val::Real; condition::Function = ==)

Get the data whose dataField, meets the condition(dataField[:,idxComp], val). It errors if the fieldname provided does not have a column idxComp.

dataCond(network::DislocationNetwork, dataField::Symbol, condField::Symbol,
    val::Real; condition::Function = ==)

Get the data from dataField that corresponds to the condField (from the same pool as dataField) that meets the condition(condField, val). If condField is multidimensional it will use CartesianIndex in its search. dataField and condField must have the same number of rows.

dataCond(network::DislocationNetwork, dataField::Symbol, condField::Symbol,
    idxComp::Integer, val::Real; condition::Function = ==)

Get the data from dataField that corresponds to the condField that meets the condition(condField[:, idxComp], val). dataField and condField must have the same number of rows.

DDD.extAngle — Method

extAngle(n::Int64)

Calculates the exterior angle of a regular polygon with n sides.

DDD.idxCond — Method

Related functions: dataCond

idxCond(network::DislocationNetwork, fieldname::Symbol,
    args...; condition::Function)

Find index/indices whose fieldname meets the condition(fieldname, args...) where condition can be any function that uses fieldname and args to make a comparison.

idxCond(data::Union{
        AbstractArray{<:Real,N1},
        AbstractArray{<:nodeType, {N2}}
    },
    val::Real; condition::Function = ==) where {N1,N2}

Find index/indices of data that meet the condition(data, val).

idxCond(network::DislocationNetwork, fieldname::Symbol, val::Real;
    condition::Function = ==)

Find index/indices of node whose fieldname, (:links, :bVec, :slipPlane, :coord, :label, :numNode, :numSeg), meets condition(fieldname, val). If fieldName is multidimensional it will return CartesianIndex.

idxCond(network::DislocationNetwork, fieldname::Symbol, idxComp::Integer,
    val::Real; condition::Function = ==)

Find index/indices of node whose fieldname meets condition(fieldname[:, idxComp], val). It errors if the fieldname provided does not have a column idxComp.

DDD.idxLabel — Method

Related functions: coordLbl

idxLabel(network::DislocationNetwork, label::Integer; condition::Function = ==)

Find indices for dislocations whose label meets condition(x, label).

DDD.inclusiveComparison — Method

inclusiveComparison(data, args...)

Compare data to a tuple, return true if it is equal to any arg, false if it is not equal to any.

DDD.intAngle — Method

intAngle(n::Int64)

Calculates the interior angle of a regular polygon with n sides.

DDD.loopDistribution — Method

loopDistribution(dist<:AbstractDistribution, n::Integer, args...; kw...)

Returns n points according to the concrete subtype of AbstractDistribution given. Overload this function with new concrete subtypes and custom distributions. This and limits! are used in translatePoints to distribute dislocations in the simulation domain.

DDD.makeConnect — Method

makeConnect(
    links::AbstractArray{Int64, N},
    label::Vector{nodeType},
    maxConnect::Integer,
) where {N}

Returns tuple of matrices connectivity and linksConnect. connectivity contains the number of other other nodes each node is connected to, up to maxConnect other nodes. It also contains the segments in which it's involved. linksConnect is the running total of the number of other nodes each node is connected to, will probably be deleted in the future. This is called from makeNetwork and makeNetwork!.

DDD.makeLoop — Method

makeLoop(
    loopType::T1,   # Type of loop.
    numSides::T2,   # Number of sides in loop, must be even.
    nodeSide::T2,   # Nodes per side.
    numLoops::T2,   # Number of loops in initial network.
    segType::T3,    # Type of segments in loop (future-proofing).
    segLen::T4,     # Lengths of half the segments in loop, why `numSides` must be even.
    slipSystem::T2, # Index of the loop's slip system.
    _slipPlane::T5, # Slip plane vector.
    _bVec::T5,      # Burgers vector.
    label::T6,      # Labels, length must be equal to `numSides * nodeSide`.
    buffer::T7,     # Buffer distance to prevent overlap.
    range::T8,      # Spatial range of `dist`, defines the space occupied by the loops in the network.
    dist::T9,       # Spatial distribution of loops in `range`.
) where {
    T1 <: AbstractDlnStr, # Dislocation structure.
    T2 <: Int64,
    T3 <: AbstractDlnSeg, # Dislocation segment type.
    T4 <: Union{
        T where {T <: Float64},
        AbstractArray{<:Float64, N} where {N}
    },
    T5 <: AbstractArray{<:Float64, N} where {N},
    T6 <: Vector{nodeType}, # Dislocation node enumerated type.
    T7 <: Float64,
    T8 <: AbstractArray{<:Float64, N} where {N},
    T9 <: AbstractDistribution, # Spatial distribution type, defines the loops' spatial distribution in the network.
}

Constructor function for DislocationLoop. See AbstractDlnStr, AbstractDlnSeg, nodeType, AbstractDistribution for further details.

DDD.makeNetwork — Function

makeNetwork(
    sources::Union{
            DislocationLoop,
            AbstractVector{<:DislocationLoop}
        }, # Dislocation structures.
    maxConnect::Integer = 4,
    memBuffer::Integer = 10, # Buffer for memory allocation. The code will allocate the total number of nodes times `memBuffer` to reduce dynamic memory allocation during runtime.
    args...;
    checkConsistency::Bool = false, # Check the consistency of the network.
    kw...,
)

Constructor for DislocationNetwork, see makeNetwork! for in-place version.

DDD.makeNetwork! — Function

makeNetwork!(
    network::DislocationNetwork,
    sources::Union{
        DislocationLoop,
        AbstractVector{<:DislocationLoop}
    },
    maxConnect::Integer = 4,
    args...;
    checkConsistency::Bool = false,
    kw...,
)

In-place constructor for DislocationNetwork, see makeNetwork for constructor.

DDD.makeSegment — Method

makeSegment(type::AbstractDlnSeg, slipPlane::Vector{T}, bVec::Vector{T}) where {T<:Float64}

Make segment depending on the segment type, see AbstractDlnSeg.

DDD.rot3D — Method

rot3D(xyz::Vector{<:Float64}, uvw::Vector{Float64}, abc::Vector{<:Float64},
    θ::Float64)

Rotate point xyz about the line with direction vector uvw that crosses the point abc by the angle θ.

DDD.saveParams — Method

Saves simulation parameters.

Base.isequal — Method

Overloaded functions for dislocation nodeType.

DDD.limits! — Method

limits!(
    lims::AbstractArray{<:Float64, N1},
    segLen::Float64,
    range::AbstractArray{<:Float64, N2},
    buffer::Float64,
) where {N1, N2}

Calculate the spatial limits a dislocation will occupy. This and loopDistribution are used in translatePoints to distribute dislocations in the simulation domain.

DDD.pushToDataFrame! — Method

Pushes data to a dataframe for saving later.

DDD.translatePoints — Method

function translatePoints(
    coord::AbstractArray{<:Float64, N1},
    lims::AbstractArray{<:Float64, N1},
    disp::AbstractArray{<:Float64, N2},
) where {N1, N2}

Translate dislocation node coordinates coord inside the spatial bounds of lims (calculated in limits!) according to the displacement disp (calculated in loopDistribution). Used to distribute sources inside a domain in makeNetwork and makeNetwork!.

DDD.@makeType — Macro

makeType(type::Any, supertype::Any)

Macro to make a type that is a subtype of supertype.

DDD.@string_as_varname_macro — Macro

string_as_varname_macro(s::AbstractString, v::Any)

Turn a string s into a variable whose name is the string s and value is v.