IO

Input

DDD.loadJSON — Method

loadJSON(filename::AbstractString)

Wrapper for JSON.parsefile(filename).

DDD.loadDislocationLoopJSON — Method

function loadDislocationLoopJSON(
    dict::Dict{T1, T2} where {T1, T2},
    slipSystem::SlipSystem,
)

Loads initial dislocation structure out of a dictionary loaded from a JSON file. Returns a variable of type DislocationLoop.

DDD.loadMaterialParametersJSON — Method

loadMaterialParametersJSON(dict::Dict{T1, T2}) where {T1, T2}

Loads material parameters out of a dictionary loaded from a JSON file. Returns a variable of type MaterialParameters.

DDD.loadIntegrationParametersJSON — Method

loadIntegrationParametersJSON(dict::Dict{T1, T2}) where {T1, T2}

Loads integration parameters out of a dictionary loaded from a JSON file. Returns a variable of type IntegrationParameters.

DDD.loadSlipSystemJSON — Method

loadSlipSystemJSON(dict::Dict{T1, T2}) where {T1, T2}

Loads slip systems out of a dictionary loaded from a JSON file. Returns a variable of type SlipSystem.

DDD.loadDislocationParametersJSON — Method

loadDislocationParametersJSON(dict::Dict{T1, T2}) where {T1, T2}

Loads dislocation parameters out of a dictionary loaded from a JSON file. Returns a variable of type DislocationParameters.

DDD.loadParametersJSON — Method

loadParametersJSON(
    fileDislocationParameters::AbstractString,
    fileMaterialParameters::AbstractString,
    fileIntegrationParameters::AbstractString,
    fileSlipSystem::AbstractString,
    fileDislocationLoop::AbstractString,
)

Loads simulation parameters out of a dictionary loaded from a JSON file. Returns a tuple of variable types (DislocationParameters, MaterialParameters, IntegrationParameters, SlipSystem, DislocationLoop) or vectors of those types.

DDD.loadNetworkJSON — Method

loadNetworkJSON(fileDislocationNetwork::AbstractString)

Loads a dislocation network from a JSON file. Returns a DislocationNetwork.

Output

DDD.saveJSON — Method

saveJSON(filename::AbstractString, args...; mode::AbstractString = "w")

Wrapper for JSON.print to a file, args are the variables or structures you want to save.