Types

Material

DDD.AbstractCrystalStruct — Type

abstract type AbstractCrystalStruct end
struct BCC <: AbstractCrystalStruct end
struct FCC <: AbstractCrystalStruct end
struct HCP <: AbstractCrystalStruct end

Crystal structure types.

DDD.MaterialP — Type

struct MaterialP{T1, T2}
    μ::T1
    μMag::T1
    ν::T1
    E::T1
    crystalStruct::T2
end

Integration

DDD.AbstractIntegrator — Type

abstract type AbstractIntegrator end
struct CustomTrapezoid <:AbstractIntegrator end

Integrator types.

DDD.IntegrationP — Type

mutable struct IntegrationP{T1, T2, T3}
    dt::T1
    tmin::T1
    tmax::T1
    method::T2
    abstol::T1
    reltol::T1
    time::T1
    step::T3
end

This structure contains the integration parameters for the simulation.

FEM

DDD.AbstractMesh — Type

abstract type AbstractMesh end

Abstract mesh type.

DDD.AbstractShapeFunction — Type

abstract type AbstractShapeFunction end
abstract type AbstractShapeFunction3D <: AbstractShapeFunction end
abstract type AbstractShapeFunction2D <: AbstractShapeFunction end
struct LinearQuadrangle3D <:AbstractShapeFunction3D end
struct LinearQuadrangle2D <:AbstractShapeFunction2D end

Shape function types.

DDD.RegularCuboidMesh — Type

struct RegularCuboidMesh{
    T1 <: AbstractArray{T3, N} where {T3, N},
    T2 <: AbstractArray{T4, N} where {T4, N},
}
    label::T1
    sizeElem::T2
    sizeMesh::T2
    stiffTensor::T2
    coord::T2
    vertices::T2
end

Cuboid mesh structure.

Dislocation

DDD.nodeType — Type

Dislocation nodes have labels that change how they are treated by the simulation. There are only certain types of nodes so these labels may only take on predefined values. Adding new nodes requires adding new entries to the enumerated type.

@enum nodeType begin
    none = 0  # Undefined node, value at initialisation.
    intMob = 1  # Internal mobile node.
    intFix = 2  # Internal fixed node.
    srfMob = 3  # Mobile surface node.
    srfFix = 4  # Fixed surface node.
    ext = 5     # External node.
end

Missing docstring.

Missing docstring for AbstractDlnSeg. Check Documenter's build log for details.

DDD.AbstractDlnStr — Type

Dislocation structures have different classifications. Prismatic loops are made up only of edge segments with the same slip system; shear loops are made up of a mixture of segment types with the same slip system; jogs and kinks are steps not contained in the slip plane.

abstract type AbstractDlnStr end
struct loopDln <: AbstractDlnStr end    # Unclassified loop.
struct loopPrism <: AbstractDlnStr end  # Prismatic loop.
struct loopShear <: AbstractDlnStr end  # Shear loop.
struct loopJog <: AbstractDlnStr end    # Jog.
struct loopKink <: AbstractDlnStr end   # Kink.

DDD.AbstractDistribution — Type

abstract type AbstractDistribution end
struct Zeros <: AbstractDistribution end
struct Rand <: AbstractDistribution end
struct Randn <: AbstractDistribution end
struct Regular <: AbstractDistribution end

Distributions for dislocation sources.

Missing docstring.

Missing docstring for AbstractMobility. Check Documenter's build log for details.

DDD.SlipSystem — Type

struct SlipSystem{T1 <: AbstractCrystalStruct, T2 <: AbstractArray{T, N} where {T, N}}
    crystalStruct::T1
    slipPlane::T2
    bVec::T2
end

Slip systems.

Dislocation-FEM

DDD.DislocationFEMCorrective — Type

mutable struct DislocationFEMCorrective{T1 <: AbstractArray{T2, N} where {T2, N}}
    uHat::T1 # U_hat
    fHat::T1 # F_hat
end

Corrective stress and force.

Composite Types