Functions

Dislocation

Segment Forces

DDD.calcSelfForce — Function

calcSelfForce(
    dlnParams::DislocationParameters,
    matParams::MaterialParameters,
    network::DislocationNetwork,
)

Calculates the self-interaction force felt by two nodes in a segment. Naturally the forces are equal and opposite to each other.

DDD.calcSegSegForce — Function

!!! Note This function is based on the SegSegForces function by A. Arsenlis et al. It is optimised for speed and reusability. It has also been locally parallelised.

It implements the analytical solution of the force between two dislocation segments. Details are found in Appendix A.1. in "Enabling Strain Hardening Simulations with Dislocation Dynamics" by A. Arsenlis et al.

At a high level this works by creating a local coordinate frame using the line directions of the dislocation segments and a vector orthogonal to them. The line integrals are then evaluated parametrically utilising this local coordinate. BibTex citation here:

@article{Arsenlis_2007, doi = {10.1088/0965-0393/15/6/001}, url = {https://doi.org/10.1088%2F0965-0393%2F15%2F6%2F001}, year = 2007, month = {jul}, publisher = {{IOP} Publishing}, volume = {15}, number = {6}, pages = {553–595}, author = {A Arsenlis and W Cai and M Tang and M Rhee and T Oppelstrup and G Hommes and T G Pierce and V V Bulatov}, title = {Enabling strain hardening simulations with dislocation dynamics}, journal = {Modelling and Simulation in Materials Science and Engineering}, abstract = {Numerical algorithms for discrete dislocation dynamics simulations are investigated for the purpose of enabling strain hardening simulations of single crystals on massively parallel computers. The algorithms investigated include the calculation of forces, the equations of motion, time integration, adaptive mesh refinement, the treatment of dislocation core reactions and the dynamic distribution of data and work on parallel computers. A simulation integrating all these algorithmic elements using the Parallel Dislocation Simulator (ParaDiS) code is performed to understand their behaviour in concert and to evaluate the overall numerical performance of dislocation dynamics simulations and their ability to accumulate percent of plastic strain.} }

DDD.calcSegForce — Function

calcSegForce(
    dlnParams::DislocationParameters,
    matParams::MaterialParameters,
    network::DislocationNetwork;
    # mesh::RegularCuboidMesh,
    # dlnFEM::DislocationFEMCorrective;
    parallel::Bool = true,
)

Dislocation Topology

Missing docstring.

Missing docstring for removeNode!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for removeConnection!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for removeLink!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for removeUpdate. Check Documenter's build log for details.

DDD.mergeNode — Function

mergeNode(network::DislocationNetwork, nodeKept::Int, nodeGone::Int)

Merges nodeGone into nodeKept. After calling this function there are no repeated entries, self-links or double links.

FEM

DDD.shapeFunction — Function

shapeFunction(shape<:AbstractShapeFunction, x, y, z)

Returns the shape functions of type typeof(shape) <: AbstractShapeFunction. If x,y,z are floats returns a vector of length N, different shape functons have different numbers of nodes. If given vectors, returns an array of size (N, length(x)).

Note

All coordinate vectors must be of equal length.

shapeFunctionDeriv are the 1st order derivatives of the shape functions.

DDD.shapeFunctionDeriv — Function

shapeFunctionDeriv(shape<:AbstractShapeFunction, x, y, z)

Returns the first order derivative of the shape functions, shapeFunction, of type typeof(shape) <: AbstractShapeFunction. If x,y,z are floats returns a 2D array of size (N, 3). If given vectors, returns a 3D array of size (N, 3, length(x)).

Note

All coordinate vectors must be of equal length.

Dislocation FEM Coupling

Missing docstring.

Missing docstring for calcPKForce. Check Documenter's build log for details.

Missing docstring.

Missing docstring for calc_σ_hat. Check Documenter's build log for details.

Miscelaneous

DDD.makeInstanceDict — Function

makeInstanceDict(valType::DataType)

Make a dictionary of enumerated variable instances. Helps in translating JSON files.

Missing docstring.

Missing docstring for translateEnum. Check Documenter's build log for details.

Missing docstring.

Missing docstring for subTypeTree. Check Documenter's build log for details.

DDD.makeTypeDict — Function

makeTypeDict(valType::DataType)

Inputs contain strings that correspond to DDD data types. This function atuomatically creates a dictionary for all concrete subtypes of a given valType.

Examples

julia> abstract type MyAbstractType end
julia> struct MyStruct1 <: MyAbstractType end
julia> struct MyStruct2 <: MyAbstractType end
julia> makeTypeDict(MyAbstractType)
Dict{String,Any} with 4 entries:
  "DDD.MyStruct1()" => MyStruct1()
  "DDD.MyStruct2()" => MyStruct2()
  "MyStruct1()"     => MyStruct1()
  "MyStruct2()"     => MyStruct2()

DDD.inclusiveComparison — Function

inclusiveComparison(data, args...)::Bool

Compare data to a tuple, return true if it is equal to any arg, false if it is not equal to any.

Examples

julia> inclusiveComparison("f", 1,4,5,"f")
true
julia> inclusiveComparison(23.246, 1.5, 4, 5, "f")
false

DDD.compStruct — Function

compStruct(arg1, arg2; verbose::Bool = false)

Function that compares values of the fields of two variables arg1 and arg2 with the same structure. If verbose = true, it will print which fields are different from each other.

Examples

julia> struct MyStruct1; x; end
julia> test1 = MyStruct1(1)
MyStruct1(1)
julia> test2 = MyStruct1(5)
MyStruct1(5)
julia> compStruct(test1, test2; verbose = true)
Structures differ in field: x.
false
julia> compStruct(1, 1; verbose = true)
true
julia> compStruct(1, [1]; verbose = true)
false

DDD.intAngle — Function

intAngle(n::Int)

Calculates the interior angle of a regular polygon with n sides.

DDD.externalAngle — Function

externalAngle(n::Int)

Calculates the exterior angle of a regular polygon with n sides.

DDD.rot3D — Function

rot3D(
    xyz::AbstractVector{T1},
    uvw::AbstractVector{T2},
    abc::AbstractVector{T3},
    θ::T4,
) where {T1, T2, T3, T4}

Rotate point xyz about the line with direction vector uvw that crosses the point abc by the angle θ. Further details found here.

Examples

julia> rot3D([1;1;1],[1;0;0],[0;0;0],π/2)
3-element Array{Float64,1}:
  1.0
 -0.9999999999999999
  1.0
julia> rot3D([1;1;1],[1;0;0],[0;0;0],-π/2)
3-element Array{Float64,1}:
1.0
1.0
-0.9999999999999999
julia> rot3D([1;1;1],[1;0;0],[0;0;0],π)
3-element Array{Float64,1}:
  1.0
 -1.0000000000000002
 -0.9999999999999999