Functions

Dislocation

FEM

DDD.shapeFunction — Function

shapeFunction(shape<:AbstractShapeFunction, x, y, z)

Returns the shape functions of type typeof(shape) <: AbstractShapeFunction. If x,y,z are floats returns a vector of length N, different shape functons have different numbers of nodes. If given vectors, returns an array of size (N, length(x)).

Note

All coordinate vectors must be of equal length.

shapeFunctionDeriv are the 1st order derivatives of the shape functions.

DDD.shapeFunctionDeriv — Function

shapeFunctionDeriv(shape<:AbstractShapeFunction, x, y, z)

Returns the first order derivative of the shape functions, shapeFunction, of type typeof(shape) <: AbstractShapeFunction. If x,y,z are floats returns a 2D array of size (N, 3). If given vectors, returns a 3D array of size (N, 3, length(x)).

Note

All coordinate vectors must be of equal length.

Dislocation FEM

DDD.calcPKForce — Function

calcPKForce(
    mesh::RegularCuboidMesh,
    dlnFEM::DislocationFEMCorrective,
    network::DislocationNetwork,
)

Calculate the Peach-Koehler force on segments.

$f = (\hat{\mathbb{\sigma}} \cdot \overrightarrow{b}) \times \overrightarrow{l}$

Missing docstring.

Missing docstring for calc_σ_hat. Check Documenter's build log for details.

Miscelaneous

DDD.makeInstanceDict — Function

makeInstanceDict(valType::DataType)

Make a dictionary of enumerated variable instances. Helps in translating JSON files.

DDD.translateEnum — Function

translateEnum(
    valType::DataType,
    dict::Dict{T1, T2},
    key::T3,
) where {T1, T2, T3}

Translates the string name of enumerated types to the actual Julia type.

DDD.subTypeTree — Function

subTypeTree(t; dict = Dict(), level = 1, cutoff = 0)

Create subtype dictionary. . Adapted from https://github.com/JuliaLang/julia/issues/24741

DDD.makeTypeDict — Function

makeTypeDict(valType::DataType)

Inputs contain strings that correspond to DDD data types. This function atuomatically creates a dictionary for all concrete subtypes of a given valType.

Examples

julia> abstract type MyAbstractType end
julia> struct MyStruct1 <: MyAbstractType end
julia> struct MyStruct2 <: MyAbstractType end
julia> makeTypeDict(MyAbstractType)
Dict{String,Any} with 4 entries:
  "DDD.MyStruct1()" => MyStruct1()
  "DDD.MyStruct2()" => MyStruct2()
  "MyStruct1()"     => MyStruct1()
  "MyStruct2()"     => MyStruct2()

DDD.inclusiveComparison — Function

inclusiveComparison(data, args...)::Bool

Compare data to a tuple, return true if it is equal to any arg, false if it is not equal to any.

Examples

julia> inclusiveComparison("f", 1,4,5,"f")
true
julia> inclusiveComparison(23.246, 1.5, 4, 5, "f")
false

DDD.compStruct — Function

compStruct(arg1, arg2; verbose::Bool = false)

Function that compares values of the fields of two variables arg1 and arg2 with the same structure. If verbose = true, it will print which fields are different from each other.

Examples

julia> struct MyStruct1; x; end
julia> test1 = MyStruct1(1)
MyStruct1(1)
julia> test2 = MyStruct1(5)
MyStruct1(5)
julia> compStruct(test1, test2; verbose = true)
Structures differ in field: x.
false
julia> compStruct(1, 1; verbose = true)
true
julia> compStruct(1, [1]; verbose = true)
false

DDD.intAngle — Function

intAngle(n::Int)

Calculates the interior angle of a regular polygon with n sides.

DDD.extAngle — Function

extAngle(n::Int)

Calculates the exterior angle of a regular polygon with n sides.

DDD.rot3D — Function

rot3D(
    xyz::AbstractVector{T1},
    uvw::AbstractVector{T2},
    abc::AbstractVector{T3},
    θ::T4,
) where {T1, T2, T3, T4}

Rotate point xyz about the line with direction vector uvw that crosses the point abc by the angle θ. Further details found here.

Examples

julia> rot3D([1;1;1],[1;0;0],[0;0;0],π/2)
3-element Array{Float64,1}:
  1.0
 -0.9999999999999999
  1.0
julia> rot3D([1;1;1],[1;0;0],[0;0;0],-π/2)
3-element Array{Float64,1}:
1.0
1.0
-0.9999999999999999
julia> rot3D([1;1;1],[1;0;0],[0;0;0],π)
3-element Array{Float64,1}:
  1.0
 -1.0000000000000002
 -0.9999999999999999

DDD.dimDot — Function

dimDot(
    x::AbstractArray{T1, N},
    y::AbstractArray{T2, N};
    dim::Int = 2,
) where {T1, T2, N}

Perform dot product along dimension dim of an array, returns a vector of dot products.

DDD.dimNorm — Function

dimNorm(x::AbstractArray{T, N}; dim::Int = 2) where {T, N}

Calculate norms along dimension dim of an array, returns a vector of norms.