IO

Input

DDD.load — Method

load(filename::AbstractString)

Wrapper for JSON.parsefile(filename).

DDD.loadDislocationLoop — Method

function loadDislocationLoop(
    dict::Dict{T1, T2} where {T1, T2},
    slipSystem::SlipSystem,
)

Loads initial dislocation structure out of a dictionary loaded from a JSON file. Returns a variable of type DislocationLoop.

DDD.loadMaterialP — Method

loadMaterialP(dict::Dict{T1, T2}) where {T1, T2}

Loads material parameters out of a dictionary loaded from a JSON file. Returns a variable of type MaterialP.

DDD.loadIntegrationP — Method

loadIntegrationP(dict::Dict{T1, T2}) where {T1, T2}

Loads integration parameters out of a dictionary loaded from a JSON file. Returns a variable of type IntegrationP.

DDD.loadSlipSystem — Method

loadSlipSystem(dict::Dict{T1, T2}) where {T1, T2}

Loads slip systems out of a dictionary loaded from a JSON file. Returns a variable of type SlipSystem.

DDD.loadDislocationP — Method

loadDislocationP(dict::Dict{T1, T2}) where {T1, T2}

Loads dislocation parameters out of a dictionary loaded from a JSON file. Returns a variable of type DislocationP.

DDD.loadParams — Method

loadParams(
    fileDislocationP::AbstractString,
    fileMaterialP::AbstractString,
    fileIntegrationP::AbstractString,
    fileSlipSystem::AbstractString,
    fileDislocationLoop::AbstractString,
)

Loads simulation parameters out of a dictionary loaded from a JSON file. Returns a tuple of variable types (DislocationP, MaterialP, IntegrationP, SlipSystem, DislocationLoop) or vectors of those types.

DDD.loadNetwork — Method

loadNetwork(fileDislocationNetwork::AbstractString)

Loads a dislocation network from a JSON file. Returns a DislocationNetwork.

Output

DDD.save — Method

save(filename::AbstractString, args...; mode::AbstractString = "w")

Wrapper for JSON.print to a file, args are the variables or structures you want to save.